Reuse the same framer across receive callbacks for a given connection
Otherwise, if the the frame is received in pieces, the new framer will
not know, and incorrectly try to read a header in the middle of a
frame.

This bug was introduced in commit
f6dbf4dfaadf47fc646a63cda502fbc2a3bb0c81 when TcpBusSide gained multiple
connections instead of one. It looks like the idea in constructing the
new framer was to stop the framer from being shared between connections.

I have placed the framer construction so that each connection has its
own, and reuses it on each receive. I have also added a unit test
that reproduces the bug.