Remove BSON dependency from RD.Transport
This is a breaking change in the TcpMessage serialization - both sides
of the connection would have to be upgraded.

Motivation
- Fewer dependencies are better. We were only using BSON to serialize
a string, we can do that ourselves.
- Removes complications around NET40 conditional compilation
- As a happy side effect, this is 2x faster for small messages

The strings themselves are encoded with UTF8 both before and after, so
problems arising from encoding are not anticipated.

I have removed 3 unit tests that seem to me to be coupled to the
implementation details.